Add a flag to stop running TestSuite on TestCase failure
This allows running tests as usual, but will stop a TestSuite at the
first TestCase failure.
This makes sense in case there are dependencies between TestCases, eg.
the first TestCase creates a resource on a REST API, and following
TestCases are doing some stuff with this resource.

I tested it locally (with and without the new flag), but did not find a
proper place in the repository to add unit tests for that feature.